refactor(Portal): rewrite js to ts
